# docs/api/openapi.yml
openapi: 3.0.3
info:
  title: Algorithmic Trading System API
  description: |
    Enterprise-grade algorithmic trading platform API providing comprehensive trading, portfolio management, 
    risk management, and market data services. Built with microservices architecture and event-driven design.
    
    ## Key Features
    - **Multi-Asset Trading**: Stocks, ETFs, Options, Futures, Forex, Crypto
    - **Real-time Market Data**: Live feeds with fallback mechanisms
    - **Advanced Risk Management**: Real-time monitoring, VaR, stress testing
    - **AI-Powered Analytics**: Strategy development and optimization
    - **High-Performance Execution**: Microsecond-level latency
    
    ## Authentication
    All API endpoints require authentication via JWT tokens or API keys.
    
    ## Rate Limiting
    - Standard endpoints: 1000 requests/minute
    - Market data endpoints: 10000 requests/minute
    - Trading endpoints: 500 requests/minute
    
    ## Error Handling
    The API uses standard HTTP status codes and returns detailed error information in JSON format.
  version: 1.0.0
  contact:
    name: Trading System API Support
    email: api-support@trading-system.com
    url: https://docs.trading-system.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: https://api.trading-system.com/v1
    description: Production server
  - url: https://staging-api.trading-system.com/v1
    description: Staging server
  - url: http://localhost:8000/v1
...
security:
  - BearerAuth: []
  - ApiKeyAuth: []
paths:
...
  /orders:
    get:
      tags: [Trading]
      summary: Get orders
      description: Retrieve user's orders with optional filtering
      operationId: getOrders
      parameters:
        - name: status
          in: query
          schema:
            type: string
            enum: [pending, filled, cancelled, rejected]
        - name: symbol
          in: query
          schema:
            type: string
        - name: limit
          in: query
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
      responses:
        '200':
          description: Orders retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Order'

    post:
      tags: [Trading]
      summary: Place order
      description: Place a new trading order
      operationId: placeOrder
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/OrderRequest'
      responses:
        '201':
          description: Order placed successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Invalid order parameters
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /orders/{order_id}:
    get:
      tags: [Trading]
      summary: Get order details
      description: Retrieve details of a specific order
      operationId: getOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order details retrieved successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '404':
          description: Order not found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

    delete:
      tags: [Trading]
      summary: Cancel order
      description: Cancel a pending order
      operationId: cancelOrder
      parameters:
        - name: order_id
          in: path
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Order cancelled successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Order'
        '400':
          description: Order cannot be cancelled
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'

  # Portfolio
...
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    OrderRequest:
      type: object
      required: [symbol, side, quantity, order_type]
      properties:
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
          minimum: 0
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        price:
          type: number
          minimum: 0
        stop_price:
          type: number
          minimum: 0
        time_in_force:
          type: string
          enum: [day, gtc, ioc, fok]
          default: day
        client_order_id:
          type: string
    Order:
      type: object
      properties:
        id:
          type: string
          format: uuid
        client_order_id:
          type: string
        symbol:
          type: string
        side:
          type: string
          enum: [buy, sell]
        quantity:
          type: number
        filled_quantity:
          type: number
        remaining_quantity:
          type: number
        order_type:
          type: string
          enum: [market, limit, stop, stop_limit]
        price:
          type: number
        stop_price:
          type: number
        status:
          type: string
          enum: [pending, filled, partially_filled, cancelled, rejected]
        time_in_force:
          type: string
          enum: [day, gtc, ioc, fok]
        created_at:
          type: string
          format: date-time
        updated_at:
          type: string
          format: date-time
        fills:
          type: array
          items:
            $ref: '#/components/schemas/Fill'
    Fill:
      type: object
      properties:
        id:
          type: string
          format: uuid
        quantity:
          type: number
        price:
          type: number
        timestamp:
          type: string
          format: date-time
        commission:
          type: number

    # Portfolio Schemas
...
    Error:
      type: object
      properties:
        error:
          type: string
        message:
          type: string
        code:
          type: integer
        timestamp:
          type: string
          format: date-time
        request_id:
          type: string
          format: uuid
...